+++ lib/Api/LoyaltyMerchantCardsApi.php
@@ -632,12 +632,13 @@ class LoyaltyMerchantCardsApi
      * @param string $scroll_expire How long it should keep the “search context” alive? 
      * @param string $scroll_id Identifier of a previous search context. 
      * @param string $preset Query presets 
+     * @param string $expand Expand fields (all or a specific one) 
      * @throws ApiException on non-2xx response
      * @return \Secuconnect\Client\Model\LoyaltyMerchantcardsList
      */
-    public function getAll($count = null, $offset = null, $fields = null, $q = null, $sort = null, $aggregate = null, $meta = null, $validate = null, $scroll_expire = null, $scroll_id = null, $preset = null)
+    public function getAll($count = null, $offset = null, $fields = null, $q = null, $sort = null, $aggregate = null, $meta = null, $validate = null, $scroll_expire = null, $scroll_id = null, $preset = null, $expand = null)
     {
-        list($response) = $this->getAllWithHttpInfo($count, $offset, $fields, $q, $sort, $aggregate, $meta, $validate, $scroll_expire, $scroll_id, $preset);
+        list($response) = $this->getAllWithHttpInfo($count, $offset, $fields, $q, $sort, $aggregate, $meta, $validate, $scroll_expire, $scroll_id, $preset, $expand);
         return $response;
     }
 
@@ -657,10 +658,11 @@ class LoyaltyMerchantCardsApi
      * @param string $scroll_expire How long it should keep the “search context” alive? 
      * @param string $scroll_id Identifier of a previous search context. 
      * @param string $preset Query presets 
+     * @param string $expand Expand fields (all or a specific one) 
      * @throws ApiException on non-2xx response
      * @return array of \Secuconnect\Client\Model\LoyaltyMerchantcardsList, HTTP status code, HTTP response headers (array of strings)
      */
-    public function getAllWithHttpInfo($count = null, $offset = null, $fields = null, $q = null, $sort = null, $aggregate = null, $meta = null, $validate = null, $scroll_expire = null, $scroll_id = null, $preset = null)
+    public function getAllWithHttpInfo($count = null, $offset = null, $fields = null, $q = null, $sort = null, $aggregate = null, $meta = null, $validate = null, $scroll_expire = null, $scroll_id = null, $preset = null, $expand = null)
     {
         // parse inputs
         $resourcePath = "/Loyalty/MerchantCards";
@@ -717,6 +719,10 @@ class LoyaltyMerchantCardsApi
         // query params
         if ($preset !== null) {
             $queryParams['preset'] = $this->apiClient->getSerializer()->toQueryValue($preset);
+        }
+        // query params
+        if ($expand !== null) {
+            $queryParams['expand'] = $this->apiClient->getSerializer()->toQueryValue($expand);
         }
 
         // for model (json/xml)