+++ CHANGELOG.md
@@ -18,6 +18,56 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Removed
 
+## [2.47.0] - 2024-04-16
+[2.47.0]:https://github.com/secuconnect/secuconnect-php-sdk/compare/2.46.0...2.47.0
+
+- *changed response type*: `PrepaidContractsApi.mappingZvt()` changed from `PrepaidMappingZvt` to `PrepaidMappingZvtResponse`, which inherit from the `PrepaidMappingZvt` (like before) or returns the `ResultBoolean` object
+- *changed element type*: `product_id` changed from `int` to `string` in `PrepaidZvtDTO`
+- *new request element*: `expand` was added to all `getAll()` endpoints
+- *change element type*: `parent` was changed from `ParentModel` to `GeneralMerchantsProductModel` in `GeneralMerchantsProductModel`
+- *change element type*: `acceptance_points` was changed from `BaseProductModel[]` to `GeneralMerchantsProductModel[]` in `GeneralMerchantsProductModel`
+
+### Added
+- `Model.PrepaidMappingZvtResponse`
+
+### Changed
+- `Api.CardprocessingInvoicesApi`
+- `Api.CardprocessingTransactionsApi`
+- `Api.GeneralContractsApi`
+- `Api.GeneralMerchantsApi`
+- `Api.GeneralStoresApi`
+- `Api.LoyaltyCardGroupsApi`
+- `Api.LoyaltyCardsApi`
+- `Api.LoyaltyCustomersApi`
+- `Api.LoyaltyMerchantCardsApi`
+- `Api.LoyaltyTransactionsApi`
+- `Api.PaymentContainersApi`
+- `Api.PaymentContractsApi`
+- `Api.PaymentCustomersApi`
+- `Api.PaymentInvoicesApi`
+- `Api.PaymentMandatesApi`
+- `Api.PaymentPayoutsApi`
+- `Api.PaymentPlansApi`
+- `Api.PaymentSubscriptionsApi`
+- `Api.PaymentTransactionsApi`
+- `Api.PrepaidContractsApi`
+- `Api.PrepaidItemsApi`
+- `Api.PrepaidReportsApi`
+- `Api.PrepaidSalesApi`
+- `Api.ServicesIdentRequestsApi`
+- `Api.ServicesIdentResultsApi`
+- `Api.SmartDevicesApi`
+- `Api.SmartRoutingsApi`
+- `Api.SmartTemplatesApi`
+- `Api.SmartTransactionsApi`
+- `Model.GeneralMerchantsProductModel`
+- `Model.PrepaidMappingZvt`
+- `Model.PrepaidMappingZvtItem`
+- `Model.PrepaidZvtDTO`
+- `Model.ResultBoolean`
+- `lib/ObjectSerializer.php`
+
+
 ## [2.46.0] - 2024-04-10
 [2.46.0]:https://github.com/secuconnect/secuconnect-php-sdk/compare/2.45.0...2.46.0